requirements.txt
gunicorn==22.0.0
Flask==3.0.3
transformers==4.43.0
torch==2.3.1
python-dotenv==1.0.1

